♻️ Card 컴포넌트 분할, 타입 지정
