+++ components/Body/Body.module.css
@@ -1,5 +1,5 @@
 .body {
   background-color: white;
-  width: 1200px;
+  width: 1060px;
   margin: 0 auto;
 }
