+++ components/Header/Navbar.module.css
@@ -3,10 +3,8 @@
   justify-content: space-between;
   align-items: center;
   flex-direction: center;
-  position: fixed;
+  /* position: fixed; */
   width: 100%;
-  height: 80px;
-  margin: 0;
   height: 150px;
 }
