+++ components/Header/Profile.module.css
@@ -24,6 +24,7 @@
   width: 128px;
   height: 54px;
   text-decoration: none;
+  border: none;
 }
 
 @media (min-width: 768px) and (max-width: 1199px) {
