+++ components/UI/Card/Card.module.css
@@ -1,0 +1,57 @@
+.card {
+  box-shadow: 0 0.5rem 2.5rem 0 rgba(0, 0, 0, 0.08);
+  border-radius: 1.5rem;
+  width: 21rem;
+  height: 20rem;
+}
+
+.card-image {
+  width: 100%;
+  height: 180px;
+}
+
+.card-image img {
+  object-fit: cover;
+  height: 100%;
+  width: 100%;
+  background-position: center;
+  border-radius: 1.5rem 1.5rem 0 0;
+  min-height: 12rem;
+}
+
+.card-content {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  margin: 20px;
+}
+
+.card-description {
+  font-weight: 400;
+  height: 3.2rem;
+  text-overflow: ellipsis;
+  overflow: hidden;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  font-size: 1.1rem;
+}
+
+.card-uploaded-when,
+.card-description,
+.card-createdAt {
+  background-color: rgba(255, 255, 255, 0.8);
+  margin-bottom: 5px;
+}
+
+@media (max-width: 1024px) {
+  .card {
+    width: calc(50% - 20px);
+  }
+}
+
+@media (max-width: 768px) {
+  .card {
+    width: calc(100% - 20px);
+  }
+}
