+++ components/UI/Card/CardList.module.css
@@ -1,66 +1,9 @@
-.card-container {
+.card-list {
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   grid-template-rows: repeat(3, 1fr);
-  justify-content: center;
-  gap: 3rem;
+  row-gap: 1rem;
   margin-bottom: 7rem;
-}
-
-.card {
-  box-shadow: 0 0.5rem 2.5rem 0 rgba(0, 0, 0, 0.08);
-  border-radius: 1.5rem;
-  width: 24rem;
-  height: 23rem;
-}
-
-.card-image {
-  width: 100%;
-}
-
-.card-image img {
-  object-fit: cover;
-  height: 100%;
-  width: 100%;
-  background-position: center;
-  border-radius: 1.5rem 1.5rem 0 0;
-  min-height: 14rem;
-  height: 10rem;
-}
-
-.card-content {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  margin: 20px;
-}
-
-.card-description {
-  height: 3.2rem;
-  text-overflow: ellipsis;
-  width: 355px;
-  overflow: hidden;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  font-size: 1.1rem;
-}
-
-.card-uploaded-when,
-.card-description,
-.card-createdAt {
-  background-color: rgba(255, 255, 255, 0.8);
-  margin-bottom: 5px;
-}
-
-@media (max-width: 1024px) {
-  .card {
-    width: calc(50% - 20px);
-  }
-}
-
-@media (max-width: 768px) {
-  .card {
-    width: calc(100% - 20px);
-  }
+  align-items: center;
+  padding-bottom: 70px;
 }
