+++ get-data/FetchFolderData.tsx
@@ -1,5 +1,12 @@
 import { useEffect, useState } from "react";
 
+export interface Data {
+  id: string;
+  imageSource: string;
+  description: string;
+  createdAt: string;
+}
+
 interface FolderData {
   folder: {
     owner: {
@@ -7,7 +14,7 @@ interface FolderData {
       name: string;
     };
     name: string;
-    links: [];
+    links: Data[];
   };
 }
 
@@ -18,11 +25,10 @@ function FetchFolderData(): FolderData | null {
     async function getFolder() {
       try {
         const response = await fetch(
-          "https://bootcamp-api.codeit.kr/api/sample/folder"
+          "https://bootcamp-api.codeit.kr/api/users/8595/folders"
         );
         const result = await response.json();
         console.log(result);
-
         setfolderData(result);
       } catch (error) {
         console.error("에러 발생!!!!!", error);
