+++ pages/_app.tsx
@@ -1,4 +1,5 @@
 // 전체에 적용할 때 _app.tsx
+
 import "../styles/global.css";
 import type { AppProps } from "next/app";
 import Footer from "@/components/Footer/Footer";
