+++ pages/_document.tsx
@@ -1,5 +1,4 @@
-
-/// 공통된 HTML 지정 가능한 파일
+// 공통된 HTML 지정 가능한 파일
 
 import { Html, Head, Main, NextScript } from "next/document";
 
@@ -12,6 +11,5 @@ export default function Document() {
         <NextScript />
       </body>
     </Html>
-
   );
 }
