+++ pages/index.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
 
 export default function Home() {
-  return <Link href="/home">home</Link>;
+  return <Link href="/folder">folder page</Link>;
 }
